+++ apps/website/src/components/icon-list-item/icon-list-item.tsx
@@ -64,7 +64,7 @@ const IconListItem = ({ icon: Icon, iconName }: IconListItemProps) => {
             onMouseLeave={handleMouseLeave}
         >
             <span className={styles.iconContainer}>
-                <Icon size="40" color="var(--foreground-normal)" />
+                <Icon size="40" color="var(--vapor-color-foreground-secondary-darker)" />
             </span>
             <Text typography="body3" color="foreground-normal" className={styles.text}>
                 {iconName}
@@ -82,19 +82,12 @@ const IconListItem = ({ icon: Icon, iconName }: IconListItemProps) => {
                         {isCopied ? (
                             <CheckCircleIcon
                                 size="16"
-                                color="var(--foreground-success-on-transparent"
+                                color="var(--vapor-color-foreground-success)"
                             />
                         ) : (
-                            <CopyIcon size="16" color="var(--foreground-hint-on-transparent" />
+                            <CopyIcon size="16" color="var(--vapor-color-foreground-contrast)" />
                         )}
-                        <Text
-                            color={
-                                isCopied
-                                    ? 'foreground-success-on-transparent'
-                                    : 'foreground-hint-on-transparent'
-                            }
-                            typography="subtitle1"
-                        >
+                        <Text foreground={isCopied ? 'success' : 'contrast'} typography="subtitle1">
                             {isCopied ? '코드 복사됨' : '코드 복사'}
                         </Text>
                     </button>