+++ packages/core/src/components/badge/badge.tsx
@@ -1,17 +1,17 @@
+'use client';
+
 import type { ComponentPropsWithoutRef } from 'react';
 import { forwardRef } from 'react';
 
+import { Primitive } from '@radix-ui/react-primitive';
 import clsx from 'clsx';
 
-import { vapor } from '~/libs/factory';
-import type { MergeRecipeVariants } from '~/libs/recipe';
 import { createSplitProps } from '~/utils/create-split-props';
 
+import type { BadgeVariants } from './badge.css';
 import * as styles from './badge.css';
 
-type BadgePrimitiveProps = Omit<ComponentPropsWithoutRef<typeof vapor.span>, 'color'>;
-type BadgeVariants = MergeRecipeVariants<typeof styles.root>;
-
+type BadgePrimitiveProps = Omit<ComponentPropsWithoutRef<typeof Primitive.span>, 'color'>;
 interface BadgeProps extends BadgePrimitiveProps, BadgeVariants {}
 
 const Badge = forwardRef<HTMLSpanElement, BadgeProps>(({ className, children, ...props }, ref) => {
@@ -22,13 +22,13 @@ const Badge = forwardRef<HTMLSpanElement, BadgeProps>(({ className, children, ..
     ]);
 
     return (
-        <vapor.span
+        <Primitive.span
             ref={ref}
             className={clsx(styles.root(variantsProps), className)}
             {...otherProps}
         >
             {children}
-        </vapor.span>
+        </Primitive.span>
     );
 });
 Badge.displayName = 'Badge';