+++ packages/core/src/components/card/card.css.ts
@@ -1,27 +1,27 @@
-import { vars } from '~/styles/contract.css';
-import { layerStyle } from '~/styles/layers.css';
+import { layerStyle } from '~/styles/utils/layer-style.css';
+import { vars } from '~/styles/vars.css';
 
-export const root = layerStyle('component', {
+export const root = layerStyle('vapor-component', {
     display: 'flex',
     flexDirection: 'column',
 
     border: `1px solid ${vars.color.border.normal}`,
     borderRadius: vars.size.borderRadius[300],
 
-    backgroundColor: vars.color.background.normal,
+    backgroundColor: vars.color.background['normal-lighter'],
 });
 
-export const header = layerStyle('component', {
+export const header = layerStyle('vapor-component', {
     borderBottom: `1px solid ${vars.color.border.normal}`,
 
     padding: `${vars.size.space[200]} ${vars.size.space[300]}`,
 });
 
-export const body = layerStyle('component', {
+export const body = layerStyle('vapor-component', {
     padding: vars.size.space[300],
 });
 
-export const footer = layerStyle('component', {
+export const footer = layerStyle('vapor-component', {
     borderTop: `1px solid ${vars.color.border.normal}`,
 
     padding: `${vars.size.space[200]} ${vars.size.space[300]}`,