+++ packages/core/src/components/h-stack/h-stack.stories.tsx
@@ -2,9 +2,10 @@ import type { ComponentProps } from 'react';
 
 import type { StoryObj } from '@storybook/react';
 
-import { vars } from '~/styles/contract.css';
+import { vars } from '~/styles/vars.css';
 
 import { HStack } from '.';
+import { Box } from '../box';
 
 export default {
     title: 'HStack',
@@ -27,11 +28,11 @@ export const Default: StoryObj<typeof HStack> = {
     render: (args) => {
         return (
             <HStack {...args}>
-                <Box size={90}>1</Box>
-                <Box size={80}>2</Box>
-                <Box size={70}>3</Box>
-                <Box size={60}>4</Box>
-                <Box size={50}>5</Box>
+                <CustomBox size={90}>1</CustomBox>
+                <CustomBox size={80}>2</CustomBox>
+                <CustomBox size={70}>3</CustomBox>
+                <CustomBox size={60}>4</CustomBox>
+                <CustomBox size={50}>5</CustomBox>
             </HStack>
         );
     },
@@ -41,29 +42,26 @@ export const TestBed: StoryObj<typeof HStack> = {
     render: (args) => {
         return (
             <HStack {...args}>
-                <Box size={90}>1</Box>
-                <Box size={80}>2</Box>
-                <Box size={70}>3</Box>
-                <Box size={60}>4</Box>
-                <Box size={50}>5</Box>
+                <CustomBox size={90}>1</CustomBox>
+                <CustomBox size={80}>2</CustomBox>
+                <CustomBox size={70}>3</CustomBox>
+                <CustomBox size={60}>4</CustomBox>
+                <CustomBox size={50}>5</CustomBox>
             </HStack>
         );
     },
 };
 
-const Box = ({ size = 50, style, ...props }: ComponentProps<'div'> & { size?: number }) => {
+const CustomBox = ({ size = 50, ...props }: ComponentProps<typeof Box> & { size?: number }) => {
     return (
-        <div
-            style={{
-                backgroundColor: vars.color.background.primary,
-                width: size,
-                height: size,
-                border: '1px solid white',
-                textAlign: 'center',
-                alignContent: 'center',
-                color: 'white',
-                ...style,
-            }}
+        <Box
+            backgroundColor="$primary"
+            width={`${size}px`}
+            height={`${size}px`}
+            border="1px solid white"
+            textAlign="center"
+            alignContent="center"
+            color="white"
             {...props}
         />
     );