+++ packages/core/src/components/icon-button/icon-button.tsx
@@ -1,18 +1,18 @@
+'use client';
+
 import type { ComponentPropsWithoutRef } from 'react';
 import { forwardRef } from 'react';
 
 import clsx from 'clsx';
 
 import { createSlot } from '~/libs/create-slot';
-import type { MergeRecipeVariants } from '~/libs/recipe';
 import { createSplitProps } from '~/utils/create-split-props';
 
 import { Button } from '../button';
+import type { IconButtonVariants } from './icon-button.css';
 import * as styles from './icon-button.css';
 
-type IconButtonVariants = MergeRecipeVariants<typeof styles.root>;
 type IconButtonPrimitiveProps = Omit<ComponentPropsWithoutRef<typeof Button>, 'stretch'>;
-
 interface IconButtonProps extends IconButtonVariants, IconButtonPrimitiveProps {
     'aria-label': string;
 }