+++ packages/core/src/styles/mixins/interactions.css.ts
@@ -2,13 +2,13 @@ import { createVar } from '@vanilla-extract/css';
 import { calc } from '@vanilla-extract/css-utils';
 import { recipe } from '@vanilla-extract/recipes';
 
-import { vars } from '../contract.css';
-import { layerStyle } from '../layers.css';
+import { layerStyle } from '../utils/layer-style.css';
+import { vars } from '../vars.css';
 
 const ratio = createVar('opacity-ratio');
 
 export const interaction = recipe({
-    base: layerStyle('component', {
+    base: layerStyle('vapor-component', {
         position: 'relative',
 
         vars: { [ratio]: '0.08' },
@@ -19,7 +19,7 @@ export const interaction = recipe({
                 top: 0,
                 left: 0,
 
-                transition: `opacity 150ms cubic-bezier(.4,0,.2,1)`,
+                transition: `opacity 150ms ease`,
                 opacity: 0,
                 border: 'none',
                 borderRadius: 'inherit',
@@ -43,12 +43,12 @@ export const interaction = recipe({
 
     variants: {
         scale: {
-            normal: layerStyle('component', { vars: { [ratio]: '0.08' } }),
-            light: layerStyle('component', { vars: { [ratio]: '0.04' } }),
+            normal: {},
+            light: layerStyle('vapor-component', { vars: { [ratio]: '0.04' } }),
         },
 
         type: {
-            default: layerStyle('component', {
+            default: layerStyle('vapor-component', {
                 selectors: {
                     '&:hover::before': { opacity: calc.multiply(ratio, 1) },
                     '&:active::before': { opacity: calc.multiply(ratio, 2) },
@@ -57,7 +57,7 @@ export const interaction = recipe({
                     },
                 },
             }),
-            form: layerStyle('component', {
+            form: layerStyle('vapor-component', {
                 transition: 'border-color 150ms cubic-bezier(.4,0,.2,1)',
 
                 selectors: {