+++ apps/website/CHANGELOG.md
@@ -1,5 +1,46 @@
 # website
 
+## 0.0.5
+
+### Patch Changes
+
+- 973fa2d: Text Input: move placeholder prop from field to root
+- ca267e9: feat: Added primary color custom functionality
+- Updated dependencies [ca267e9]
+    - @vapor-ui/core@0.3.1
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [e139a7f]
+- Updated dependencies [03655b1]
+- Updated dependencies [488c614]
+    - @vapor-ui/core@0.3.0
+
+## 0.0.3
+
+### Patch Changes
+
+- 6749d80: fix(createThemeConfig): support RSC
+- Updated dependencies [6749d80]
+    - @vapor-ui/core@0.2.1
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [55f2f42]
+- Updated dependencies [c5cd0fc]
+- Updated dependencies [4725a73]
+- Updated dependencies [9f96e2c]
+- Updated dependencies [b59dd77]
+- Updated dependencies [267a998]
+- Updated dependencies [a0c1ff0]
+- Updated dependencies [d7c2714]
+- Updated dependencies [d7c2714]
+    - @vapor-ui/core@0.2.0
+
 ## 0.0.1
 
 ### Patch Changes